refactored flask application to be modular using blueprints. Removed "old" DBManager.py module.
