--- tests/schema.sql
@@ -1,6 +1,0 @@
-CREATE TABLE User (
-        id INTEGER PRIMARY KEY AUTOINCREMENT,
-        name TEXT NOT NULL,
-        email TEXT NOT NULL UNIQUE,
-        pw TEXT NOT NULL
-    );
